% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.r
\name{add_logo}
\alias{add_logo}
\title{Function to add a logo to a ggplot that uses theme_pub}
\usage{
add_logo(g = NULL, org = "yale")
}
\arguments{
\item{g}{A ggplot created with theme_pub or an organization-specific theme}

\item{org}{name of the org. Must match the name of the .png in the logo folder.}
}
\description{
Function to add a logo to a ggplot that uses theme_pub
}
