% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgcolors.r
\docType{data}
\name{cmu.pal}
\alias{cmu.pal}
\title{A color palette for CMU used in `orgthemes`}
\format{
An object of class \code{character} of length 9.
}
\usage{
cmu.pal
}
\description{
The colors are listed in the order we typically want to use them
}
\keyword{datasets}
