% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.ggplot.defaults.r
\name{restore.ggplot.defaults}
\alias{restore.ggplot.defaults}
\title{Restore ggplot defaults}
\usage{
restore.ggplot.defaults()
}
\description{
A function for restoring ggplot defaults that were changed by the themesn package.
}
\examples{

## Load required packages
library(tidyverse)
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + geom_point() ## default discrete colors (red, green, blue)

## themesn discrete colors (snred, snblue, snmediumgray)
library(themesn)
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + geom_point() + theme_sn()

## Still themesn discrete colors (snred, snblue, snmediumgray), even though theme_sn isn't used
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + geom_point()

## Restore defaults. Now the original colors are used again.
restore.ggplot.defaults()
ggplot(mtcars, aes(x=wt, y=mpg, color=as.factor(cyl))) + geom_point()
ggplot(mtcars, aes(x=wt, y=mpg, color=          cyl )) + geom_point()
}
