% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.lims.r
\name{square.lims}
\alias{square.lims}
\title{Given data, create square limits based on that data}
\usage{
square.lims(data = NULL, symmetric = F)
}
\arguments{
\item{data}{Data to base the limits on}

\item{symmetric}{If TRUE, then the limits will be symmetric about the origin.  Default is FALSE.}
}
\description{
Create square limits for scatter plots, meaning x and y have the same limits.
}
