% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgcolors.r
\name{theme_yale}
\alias{theme_yale}
\title{Yale ggplot theme called theme_yale.}
\usage{
theme_yale(
  type = "line",
  base_size = 36/3,
  base_family = "sans",
  base_line_size = base_size * 0.35/36 * 3,
  base_rect_size = base_size * 0.35/36,
  facet = F
)
}
\arguments{
\item{type}{Same as in `theme_pub`}

\item{base_size}{Same as in `theme_pub`}

\item{base_family}{Same as in `theme_pub`}

\item{base_line_size}{Same as in `theme_pub`}

\item{base_rect_size}{Same as in `theme_pub`}

\item{facet}{Same as in `theme_pub`}
}
\description{
Same as theme_pub but with the Yale-specific colors
}
