% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgcolors.r
\docType{data}
\name{yale.pal}
\alias{yale.pal}
\title{A color palette for Yale used in `orgthemes`}
\format{
An object of class \code{character} of length 8.
}
\usage{
yale.pal
}
\description{
The colors are listed in the order we typically want to use them
}
\keyword{datasets}
