% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgcolors.r
\docType{data}
\name{yalegreen}
\alias{yalegreen}
\title{A color used in `orgthemes`}
\format{
An object of class \code{character} of length 1.
}
\usage{
yalegreen
}
\description{
A color used in `orgthemes`
}
\keyword{datasets}
